fix: fixed the bug of the findIngredientByCriteria due to the new priceList implementation

+++ src/main/java/org/radiant_wizard/db/migration/dish_table_creation.sql
@@ -1,5 +1,5 @@
 create table dishes(
     dish_id int primary key,
     dish_name varchar(50) not null,
-    dish_price int
+    dish_price NUMERIC(10,2)
 );
